+++ R/compare_means.R
@@ -338,11 +338,11 @@ compare_means <- function(formula, data, method = "wilcox.test",
 
   pvalues <- suppressWarnings(do.call(test, test.opts)$p.value) %>%
     as.data.frame()
-  group1 <- group2 <- p <- NULL
-  pvalues$group2 <- rownames(pvalues)
+  ..group1.. <- ..group2.. <- p <- NULL
+  pvalues$..group2.. <- rownames(pvalues)
   pvalues <- pvalues %>%
-    tidyr::gather(key = "group1", value = "p", -group2) %>%
-    dplyr::select(group1, group2, p) %>%
+    tidyr::gather(key = "..group1..", value = "p", -..group2..) %>%
+    dplyr::select(group1 = ..group1.., group2 = ..group2.., p) %>%
     dplyr::filter(!is.na(p))
   pvalues
 }